Find Element Widget Creation
Here are three files for the Find Element Widget. The uifind.py file has all of the background compatibility programmed for the UI and is ready for hardware specific programming and file interaction. The other two files are the latest ui file from Qt designer and that same file again but converted to python.

Co-Authored-By: Axel Montano <73506407+Biowire200@users.noreply.github.com>